Move blockSignals to silx.gui.utils
