+++ silx/gui/utils/__init__.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 # /*##########################################################################
 #
-# Copyright (c) 2018 European Synchrotron Radiation Facility
+# Copyright (c) 2018-2019 European Synchrotron Radiation Facility
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -27,3 +27,22 @@
 __authors__ = ["T. Vincent"]
 __license__ = "MIT"
 __date__ = "09/03/2018"
+
+
+import contextlib as _contextlib
+
+
+@_contextlib.contextmanager
+def blockSignals(*objs):
+    """Context manager blocking signals of QObjects.
+
+    It restores previous state when leaving.
+
+    :param qt.QObject objs: QObjects for which to block signals
+    """
+    blocked = [(obj, obj.blockSignals(True)) for obj in objs]
+    try:
+        yield
+    finally:
+        for obj, previous in blocked:
+            obj.blockSignals(previous)
